Split algorithms module by responsibility
